refactor: extract build script template and whitelist generated dirs
- Move buildScriptContent into build-script-template.ts module
- Replace blacklist (node_modules, dist, .apical-ts-build) in file
  counting with whitelist of generated directories (schemas, client,
  server, routes)
